[Misc]: split fuzz targets into 2; one for yaml, one for dsl
